Restrict ToJson SMT to JSON-ish Schemas, replace with String Schema
As we are creating a JSON String, it only makes sense to support Connect
Schema types that can be converted to a valid JSON document:
1. Array  - [1, 2, 3]
2. Map    - { "key" : "value" }
3. Struct - { "key" : "value" }

If one of the above types are not detected, a `ConnectException` will be
thrown explaining what types are supported.

If the value is a primitive type, for example a String `"my-String"`, it
does not make sense to try and create a JSON String from the value as
there is no array or key:value pair to create the valid JSON document.

Also, the underlying `ObjectMapper` will simply create a String for the
value which is not a valid JSON document.

---

When we transform the value of the SourceRecord/SinkRecord, the schema
will be changed from a Array/Map/Struct to a `String` Schema type which
relates to the JSON String that was created.

---

Unit tests have been added to test the conversion from all the core
types defined in `org.apache.kafka.connect.data.Schema.Type`:
- INT8
- INT16
- INT32
- INT64
- FLOAT32
- FLOAT64
- BOOLEAN
- STRING
- BYTES
- ARRAY
- MAP
- STRUCT

the fields with the Schema types above are all wrapped in a Struct.

The logical types also have unit tests to verify their conversion:
- org.apache.kafka.connect.data.Decimal
- org.apache.kafka.connect.data.Date
- org.apache.kafka.connect.data.Time
- org.apache.kafka.connect.data.Timestamp

A Parameterized test class has been added to check that adding a non
JSON-ish Schema type for the record value will throw an exception:
- https://github.com/junit-team/junit4/wiki/Parameterized-tests

Some unit tests have been added to get coverage on the no-op methods:
- configure
- config
- close

---

See more at
- https://github.com/apache/kafka/blob/trunk/connect/api/src/main/java/org/apache/kafka/connect/data/Schema.java
- https://github.com/apache/kafka/blob/trunk/connect/api/src/main/java/org/apache/kafka/connect/data/ConnectSchema.java